Removing use of .env file. IMPORTANT to read.
This was an artifact of an old dev env where that file was used
with foreman locally, aka no Docker in the picture.
Now, we just directly point to the the .env-docker file for the
dev env vars when bringing up the containers.

If someone does decide use a .env file to change what ENV vars
are available, def don't build that into the image! All sorts of
weird ENV var behavior can happen

IMPORTANT NOTE: after this commit, all devs should delete the
generated .env file at the app root that would have been there
before this commit, otherwise changes to .env-docker (which
controls our dev env) may be overwritten (and it will be confusing why)
-- trust me.