Allow both bebraven.org and braven.org to iframe stuff.
We have a resume and cover letter bank in the portal that iframe's
a page here on the Join server. We have two portals, one hosted at
booster.braven.org and one at portal.bebraven.org that need to be
able to iframe the resume/cover letter page. Also, soon we will be
migrating portal.bebraven.org to portal.braven.org. This commit
allows those iframes to work.

Test Plan:
- Load these two pages in prodution and make sure the iframe works:
  https://portal.bebraven.org/courses/1/pages/resume-and-cover-letter-bank
  https://booster.braven.org/courses/1/pages/resume-and-cover-letter-bank
- Make sure it also works in the dev env.
- If you get really ambitious try to iframe it on other domains in production
  and see that it doesn't work.
- If you get really, really ambitious try it out on older browsers and see what happens.